JavaMethodOverloadingExample
In this project, we show how method overloading works in Java. Method overloading is a technique that allows you to define multiple methods with the same name and call them with different parameters. This allows methods to perform the same task with different data types or different numbers of parameters.